refactoring. - current container is stored to context var - removed container name and storing all containers - make_default is renamed to use_container - Resolvers are moved to resolvers module
